+++ python/src/clier/learn/__init__.py
@@ -2,6 +2,6 @@
 Command-line help system for displaying markdown-based help topics.
 """
 
-from .help_system import HelpSystem, HelpGroup, HelpConfig, create_help_command
+from clier.learn.learn import HelpSystem, HelpGroup, HelpConfig, create_help_command
 
 __all__ = ["HelpSystem", "HelpGroup", "HelpConfig", "create_help_command"]
